+++ .gitignore
@@ -5,3 +5,4 @@ obj
 packages
 csj/csj.exe
 csj/csj.pdb
+*.user
