+++ AsyncPoco.Tests/App.config
@@ -3,7 +3,7 @@
 	<connectionStrings>
 		<add name="mysql" connectionString="server=localhost;database=asyncpoco;user id=asyncpoco;password=asyncpoco;Allow User Variables=true" providerName="MySql.Data.MySqlClient" />
 		<add name="sqlserver" connectionString="Data Source=localhost;Initial Catalog=AsyncPocoTest;Integrated Security=True" providerName="System.Data.SqlClient" />
-		<add name="sqlserverce" connectionString="Data Source=petapoco.sdf" providerName="System.Data.SqlServerCe" />
+		<add name="sqlserverce" connectionString="Data Source=|DataDirectory|\petapoco.sdf" providerName="System.Data.SqlServerCe" />
 		<add name="postgresql" connectionString="Server=127.0.0.1;User id=postgres;password=password01;Database=postgres;" providerName="Npgsql" />
 	</connectionStrings>
