+++ AsyncPoco.Tests/packages.config
@@ -3,5 +3,6 @@
   <package id="Microsoft.SqlServer.Compact" version="4.0.8876.1" targetFramework="net45" />
   <package id="MySql.Data" version="6.9.9" targetFramework="net45" />
   <package id="Npgsql" version="2.2.7" targetFramework="net45" />
-  <package id="PetaTest" version="0.1.0" />
+  <package id="NUnit" version="3.7.1" targetFramework="net45" />
+  <package id="NUnitLite" version="3.7.2" targetFramework="net45" />
 </packages>
