+++ AsyncPoco.Tests/poco.cs
@@ -1,13 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using AsyncPoco;
 
 namespace AsyncPoco.Tests
 {
 
-	enum State
+    enum State
 	{
 		Yes,
 		No,
