+++ AsyncPoco.sln
@@ -1,6 +1,8 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
+# Visual Studio 15
+VisualStudioVersion = 15.0.26430.15
+MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AsyncPoco", "AsyncPoco\AsyncPoco.csproj", "{33699753-151D-4100-8F12-9AE1AEB5C5BD}"
 	ProjectSection(ProjectDependencies) = postProject
 		{53F07078-98EA-4494-9829-74245992A7F1} = {53F07078-98EA-4494-9829-74245992A7F1}
